+++ projects/rocprofiler-compute/src/utils/utils_analysis.py
@@ -810,7 +810,7 @@ def process_rocpd_csv(df: pd.DataFrame) -> pd.DataFrame:
     # Rank GPU IDs, map lowest number to 0, next to 1, etc.
     df["GPU_ID"] = df["GPU_ID"].rank(method="dense").astype(int) - 1
     # Reset dispatch IDs
-    df["Dispatch_ID"] = range(len(df))
+    df["Dispatch_ID"] = range(1, len(df) + 1)
     return df